refactor: :recycle: Move playground card to its own file
